Fix #45: Wizard buttons crash with NameError, hidden by file-wide noqa
Two live crashes, both masked by the wizards' file-wide F821 noqa headers:

- Decision wizard "+ Effect Picker" (both call sites) invoked
  _open_effect_picker, never defined or imported anywhere. Extract the
  event wizard's own _open_effect_picker into _shared.py as
  open_effect_picker(parent, target_text, on_insert=None) and have both
  wizards call the shared implementation instead of decision.py growing
  a second copy.
- The event wizard's scripted-loc save step tested "v_id" in dir() (always
  False — v_id is never bound in that closure) then unconditionally called
  through callable(lambda: _get_output_text()) (always True, since a lambda
  is always callable), crashing any save with a scripted-loc edit target
  set. Replaced with build_event_scripted_loc_blocks(), a pure function
  over the already-in-scope `events` list, matching how the rest of the
  save step already iterates it.
